# Make encodeUtf8BuilderEscaped faster