`aeson` benchmarks signalled (https://github.com/haskell/aeson/issues/893#issuecomment-986215880) that `encodeUtf8BuilderEscaped` got slower after UTF8 transition. This PR makes it 2-3x faster:

```
All
  EncodeUtf8
    non-ASCII
      Text/encodeUtf8BuilderEscaped: OK (3.96s)
        1.90 ms ±  53 μs, 46% faster than baseline
    ASCII
      Text/encodeUtf8BuilderEscaped: OK (10.32s)
        629  μs ± 9.0 μs, 70% faster than baseline
```